requirements.txt
asgiref==3.2.10
Django==3.0.8
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
